packages/airm/pyproject.toml
# Copyright © Advanced Micro Devices, Inc., or its affiliates.
#
# SPDX-License-Identifier: MIT

[build-system]
requires = ["setuptools"]

[project]
name = "airm"
description = "Common functionality and schema for AMD Resource Manager API and Dispatcher"
version = "0.2.3"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Silo AMD", email = "info@silo.ai" }]
maintainers = [{ name = "Silo AMD", email = "info@silo.ai" }]
classifiers = ["Programming Language :: Python"]
dependencies = [
    "aio-pika>=9.5.5",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["airm*"]

[tool.pytest.ini_options]
doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"
log_cli = "0"
log_cli_level = "INFO"

[tool.mypy]
strict = false
files = ["src/**/*.py"]
ignore_missing_imports = true
pretty = true
show_column_numbers = true

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "docker>=7.1.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-docker>=3.2.1",
    "pytest-env>=1.1.5",
    "pytest-httpx>=0.35.0",
    "responses>=0.25.7",
    "uvicorn>=0.34.0",
]
